refactor(construction): split slotOperations into focused modules
Split 496-line monolithic module into 5 focused subdirectory modules:
- types.ts (24 lines): SlotAssignment, SlotOperationResult interfaces
- queries.ts (180 lines): Fixed/available slot indices, unhittable detection
- filling.ts (152 lines): Fill algorithm, unallocated unhittables
- assignment.ts (143 lines): Compact and sort algorithms
- index.ts (22 lines): Barrel export

Original file reduced to facade (1 line) preserving all public APIs.
All modules under 400 lines. Zero breaking changes to consumers.
