--- src/utils/construction/armorCalculations/index.ts
@@ -1,4 +1,0 @@
-export * from './types';
-export * from './limits';
-export * from './weight';
-export * from './allocation';
